extract CombatantsAreSame out into whenever core
